python/requirements.txt
mysql-connector-python
pymemcache
elasticsearch
kafka-python
loguru
pandas
matplotlib
dataframe-image
numpy
seaborn